Implement Sonification prototype
Sonification is implemented as an Abstract Base Class with methods
that would be called at different stages of the sonification to obtain
the OSC messages that will be used to instruct the server.

The user of the framework is expected to return a list of messages. This
means that the user will not be able to control the timing (as he could
do with a bundle). The (data playback) timing will all be handle by the
framework.
Moreover the user will implement only one function, that will be used
for all kind of sonification: real-time, real-time with recorded data
and NRT.

framework.ipynb shows a rudimentary example of how the main loop of
data player would look like. Still there are some issues to decide