Add NRT sonification to framework notebook
The implementation of NRT sonification loop using the Sonification class
demonstrates a mechanism to have different kinds of sonifications and
at the same time having the user code the sonification logic just once.
This mechanism is not complete, as we can see from the fact that we
have to instantiate the synth specifically for NRT sonification.

Explicit allocation of nodes IDs was using to remove hardcoding of node
IDs.

The notebook was cleaned before being committed.