Refactor NRT sonification with multiclient support
NRT sonification adds the default group of the client to the NRT-server.
This allows the user to use the context of the default server (e.g. for
implicit or explicit ID allocation) also for NRT sonification. This way,
the NRT sonification feature is "for free".

The concrete subclass of Sonification has been adapted to the new
expected API. Now it returns bundlers objects.