Refactor Sonification using descriptors
Parameter and its subclasses allow the user to specify, in a declarative
way, which are the parameters and which widgets should represent them in
a notebook.

Moreover, parameters accept to be updated only when the sonification
class is executing one of its functions, so that parameters can't change
in the context of the same timestamp.