Complete basic DataPlayer widget
This commit completes the basic widget of the DataPlayer. Basic
functionalities are there but a deep refactoring is needed for a proper
implementation. The limits on the interaction imposed by
ipywidgets (e.g. callback on updates) forced some changes on the logic
of the DataPlayer. It would be desirable to separate the logic, maybe
implementing a MVC achitecture.