Add live feature display for RTDataPlayer
The same object can also be used with DataPlayer.

LiveFeatureDisplay stores one list for every key observed. Elements can
be added to those lists using the feed method. Their content will be
displayed at a rate decided by the fps argument of the show method: at
regular intervals a function to update the plot will be called. This has
the advantage that we are using a standard matplotlib mechanism and we
don't have to handle threading on our own.

The displaying function does not modify any data structure, hence we
don't really need synchronization mechanisms such as locks.

We may want to refactor everything not to clear the axes and redisplay
everything all the time.