Run initialize() only when setting sonification
The initialize method is run only when setting the sonification. In
theory, it is not necessary that the _initialize() method would return a
bundle, but doing so provides a consistent API to the user. The
internals of the framework could also unpack the bundle and send it as
separate asynchronous messages. Sending a bundle with the specified
latency, should not have any negative side.