Refactor export and initialization
The initialization is divided into different stages. __init__ pass its
parameters to init_parameters.

The init_bundle stores the bundle that is sent at initialization time to
initialize the server. This is done to support the operations in export.
Somehow allocating buffers multiple times generate an error:
        Buffer UGen: no buffer data

We may want to consider to refactor everything later on.