Use time-based video frame seek in data player
This is done to support cases in which the video has a different frame
rate compared to the data, independent or irregular (using frame times).

The API of the video players was simplified eliminating every
possibility of getting an answer. The feedback will be displayed in the
output directly by the backend process.
Probably we would want to change this later on.