Remove output widgets from widget views
Output widget don't work well with multi-threading and cause lots of
problems. Output widgets just seems not to be thought for such a
context. The suggested solution of the official docs suggests to call
explicitly methods such as append_stdout on the output widget, but we
wouldn't like the user to be exposed to this; we would like the user to
be able the standard print function from inside the sonification class.

Instead the user can now re-execute the notebook cell containing the
display statements to clear the output of the cell. The only thing that
could cause problems is the fact that the output thread will attempt in
writing its output in the current cell.

If in the future we will use processes instead of threads, we will not
have this problem, as processes will write in the cell they were started
from.