Refactor to make streams return numpy arrays
This increases the performance in the creation of the Series object and
allows the user to specify the type of the numpy array, giving the user
direct control over the type.

numpy contains functions to obtain conversions from numpy types to
ctypes, which could be useful when using multiprocessing.Array to
implement a RTDataPlayerMulti using multiprocessing.