Remove data copy when concatenating slots
When using multi-threading, we don't need to copy the data, as when new
data is available, it will be simply assigned to the slot (without
overwriting the data). The data will not be garbage collected if we
already obtained a reference to it.

When using multi-processing, we already copy the data when we create the
numpy array. This part of code will most likely be refactored anyway.