Add real-time preprocessing support
This is implemented through an abstract base class called Preprocessor.
This class is used as a builder for its instances. This is better than
using a simple callback function, because it encourages the user to
consistently handle contextual data (e.g. we may need to access the
previous data sample to perform the preprocessing). A fresh instance of
the class is created every time the real-time data players start
listening.

This is a prototype. Maybe it would be a good idea to use instead a
builder pattern, so that Preprocessors can have parametric arguments.